# public/swagger-specs/services/transaction-scheduler/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PegaSys Orchestrate Transaction API. Enables dynamic management of transactions",
        "title": "Transaction Scheduler API",
        "contact": {
            "name": "Contact PegaSys Orchestrate",
            "url": "https://pegasys.tech/contact/",
            "email": "support@pegasys.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get a list of filtered transaction requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Search transaction requests by provided filters",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "List of idempotency keys",
                        "name": "idempotency_keys",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transaction requests found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/types.TransactionResponse"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "params": {
                                                "$ref": "#/definitions/types.ETHTransactionParams"
                                            },
                                            "schedule": {
                                                "$ref": "#/definitions/types.ScheduleResponse"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter in the request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/deploy-contract": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new contract deployment request\nThe transaction can be private (Tessera, Orion).\nThe transaction can be a One Time Key transaction in 0 gas private networks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Creates and sends a new contract deployment",
                "parameters": [
                    {
                        "description": "Deployment transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.DeployContractRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.DeployContractParams"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created deployment transaction request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.TransactionResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.ETHTransactionParams"
                                        },
                                        "schedule": {
                                            "$ref": "#/definitions/types.ScheduleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new smart contract transaction request\nThe transaction can be private (Tessera, Orion).\nThe transaction can be a One Time Key transaction in 0 gas private networks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Creates and sends a new contract transaction",
                "parameters": [
                    {
                        "description": "Contract transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.SendTransactionRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.TransactionParams"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created contract transaction request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.TransactionResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.ETHTransactionParams"
                                        },
                                        "schedule": {
                                            "$ref": "#/definitions/types.ScheduleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/send-raw": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new raw transaction request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Creates and sends a raw transaction",
                "parameters": [
                    {
                        "description": "Raw transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.RawTransactionRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.RawTransactionParams"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created raw transaction request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.TransactionResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.ETHTransactionParams"
                                        },
                                        "schedule": {
                                            "$ref": "#/definitions/types.ScheduleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/transfer": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new transfer request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Creates and sends a transfer transaction",
                "parameters": [
                    {
                        "description": "Transfer transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.TransferRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.TransferParams"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created transfer transaction request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.TransactionResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.ETHTransactionParams"
                                        },
                                        "schedule": {
                                            "$ref": "#/definitions/types.ScheduleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Fetch a single transaction request by uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Fetch a transaction request by uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of the transaction request",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction request found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.TransactionResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "$ref": "#/definitions/types.ETHTransactionParams"
                                        },
                                        "schedule": {
                                            "$ref": "#/definitions/types.ScheduleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Transaction request not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "types.DeployContractRequest": {
            "type": "object",
            "required": [
                "chain",
                "params"
            ],
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "myChain"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "params": {
                    "type": "object",
                    "$ref": "#/definitions/types.DeployContractParams"
                }
            }
        },
...
        "types.ScheduleResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.JobResponse"
                    }
                },
                "tenantID": {
                    "type": "string",
                    "example": "tenant_id"
                },
                "uuid": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
        "types.TransactionResponse": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "myChain"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "idempotencyKey": {
                    "type": "string",
                    "example": "myIdempotencyKey"
                },
                "params": {
                    "type": "object",
                    "$ref": "#/definitions/types.ETHTransactionParams"
                },
                "schedule": {
                    "type": "object",
                    "$ref": "#/definitions/types.ScheduleResponse"
                },
                "uuid": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}